expand autocomplete request tests, remove some noise from the suggestion struct, ensure handle new property correctly
