+++ DuckDuckGo/AutocompleteRequest.swift
@@ -55,14 +55,15 @@ class AutocompleteRequest {
         return entries.compactMap {
             guard let phrase = $0.phrase else { return nil }
 
-            if let nav = $0.nav {
-                // We definitely have a nav indication so use it
-                let url = nav ? URL(string: "http://\(phrase)") : nil
-                return Suggestion(source: .remote, type: "phrase", suggestion: phrase, url: url)
+            if let isNav = $0.nav {
+                // We definitely have a nav indication so use it. Phrase should be a fully qualified URL.
+                //  Assume HTTP and that we'll auto-upgrade if needed.
+                let url = isNav ? URL(string: "http://\(phrase)") : nil
+                return Suggestion(source: .remote, suggestion: phrase, url: url)
             } else {
                 // We need to infer nav based on the phrase to maintain previous behaviour (ie treat phrase that look like URLs like URLs)
                 let url = URL.webUrl(fromText: phrase)
-                return Suggestion(source: .remote, type: "phrase", suggestion: phrase, url: url)
+                return Suggestion(source: .remote, suggestion: phrase, url: url)
             }
         }
     }
