+++ DuckDuckGo/AutocompleteViewController.swift
@@ -157,7 +157,7 @@ class AutocompleteViewController: UIViewController {
             let matches = strongSelf.bookmarksSearch.search(query: query)
             let notQueryMatches = matches.filter { $0.url.absoluteString != query }
             let filteredMatches = notQueryMatches.filter { $0.displayTitle != nil }.prefix(Constants.maxLocalItems)
-            let localSuggestions = filteredMatches.map { Suggestion(type: "", suggestion: $0.displayTitle!, url: $0.url)}
+            let localSuggestions = filteredMatches.map { Suggestion(source: .local, suggestion: $0.displayTitle!, url: $0.url)}
             
             guard let suggestions = suggestions, error == nil else {
                 os_log("%s", log: generalLog, type: .debug, error?.localizedDescription ?? "Failed to retrieve suggestions")
